requirements.txt
flask-jwt-extended
flask-mail==0.9.0
flask-cors
flask-sqlalchemy
pymysql
pytest-flask
validate_email
python-dotenv
pytest-cov
